# contracts/openapi/cloudapp.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "properties": {
          "confirmNewPassword": {
            "type": "string"
          },
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateUserRequest": {
        "properties": {
          "confirmPassword": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "CsrfToken": {
        "properties": {
          "headerName": {
            "type": "string"
          },
          "parameterName": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "File": {
        "properties": {
          "contentType": {
            "type": "string"
          },
          "fileId": {
            "format": "int64",
            "type": "integer"
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "userid": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ],
        "type": "object"
      },
...
      "Note": {
        "properties": {
          "description": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "userid": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UpdateUserRolesRequest": {
        "properties": {
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "User": {
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer token": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "DEMO",
      "url": "https://exxxxxxxample.vercel.app/"
    },
    "description": "Documentation Cloudapp API v0.01: Bearer token is received via user-controller /api/user/login endpoint",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "Cloudapp API",
    "version": "v0.01"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/file/user/{username}": {
      "get": {
        "operationId": "getNotes_1",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/File"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "file-controller"
        ]
      }
    },
...
    "/note/user/{username}": {
      "get": {
        "operationId": "getNotes",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Note"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "note-controller"
        ]
      }
    },
...
    "/user/admin/roles": {
      "post": {
        "operationId": "updateUserRoles",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRolesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "user-controller"
        ]
      }
    },
...
    "/user/csrf-token": {
      "get": {
        "operationId": "getCsrfToken",
        "parameters": [
          {
            "in": "query",
            "name": "csrfToken",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CsrfToken"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "user-controller"
        ]
      }
    },
    "/user/id/{id}": {
      "get": {
        "operationId": "findById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "user-controller"
        ]
      }
    },
    "/user/user-change-password": {
      "post": {
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "user-controller"
        ]
      }
    },
    "/user/user-login": {
      "post": {
        "operationId": "authenticateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "user-controller"
        ]
      }
    },
...
    "/user/user-register": {
      "post": {
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "user-controller"
        ]
      }
    },
    "/user/{username}": {
      "get": {
        "operationId": "findByUserName",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "user-controller"
        ]
      }
    }
  },
  "security": [
    {
      "Bearer token": []
    }
  ]
}